requirements.txt
absl-py==1.2.0
asgiref==3.5.2
asn1crypto==1.5.1
astunparse==1.6.3
botocore==1.27.72
cachetools==3.1.1
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
cryptography==2.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.1
django-allauth==0.51.0
django-widget-tweaks==1.4.12
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
gunicorn==20.1.0
h5py==3.7.0
idna==2.8
instance==0.0.0.0.0.0a0.dev20210120
iso8601==1.0.2
jmespath==1.0.1
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mock==4.0.3
numpy==1.23.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
postgres==4.0
protobuf==3.19.4
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
skill==0.0.6
sqlparse==0.4.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.2
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
typing_extensions==4.3.0
urllib3==1.25.11
urwid==2.1.2
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
